Student app completed, and explain an issue
I face quasi the same issue or related with it that in the previous commit, I solved it by using "python manage.py migrate --fake faculty_member 0002_rename_paragrapevent_paragraphevent_paragrapharea_and_more
" command because the issue is that when I attempt to use the "python manage.py migrate" it appear that the table is already exists, that happened I think because I reset the database.

there is a second method to resolve the issue by delete the existing table directory from the database.

NOTE : Each of the to options to resolve the issue need to backup the database to not losses the important data, but in our case there is no data in the data base : ).  ..Thank you for reading this.